AF-258 Add resource access filter.
