+++ src/main/java/com/aline/core/security/filter/JwtTokenFilter.java
@@ -1,17 +1,13 @@
 package com.aline.core.security.filter;
 
-import com.aline.core.config.DisableSecurityConfig;
 import com.aline.core.exception.UnauthorizedException;
 import com.aline.core.security.config.JwtConfig;
 import com.aline.core.security.model.JwtToken;
-import com.aline.core.security.model.UserAuthToken;
 import io.jsonwebtoken.JwtException;
 import lombok.RequiredArgsConstructor;
-import lombok.val;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpHeaders;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 
@@ -23,16 +19,34 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.Optional;
 
+/**
+ * The JwtRequestFilter is an abstract class
+ * that verifies the JWT token and provides and
+ * filters based on the validity of the token.
+ * Extra logic may be implemented that is based on the
+ * token.
+ */
 @Component
-@RequiredArgsConstructor
-@ConditionalOnMissingBean(DisableSecurityConfig.class)
-public class JwtTokenFilter extends OncePerRequestFilter {
+public abstract class JwtTokenFilter extends OncePerRequestFilter {
 
-    private final JwtConfig jwtConfig;
-    private final SecretKey jwtSecretKey;
+    private JwtConfig jwtConfig;
+    private SecretKey jwtSecretKey;
+
+    @Autowired
+    public void setJwtConfig(JwtConfig jwtConfig) {
+        this.jwtConfig = jwtConfig;
+    }
+
+    @Autowired
+    public void setJwtSecretKey(SecretKey jwtSecretKey) {
+        this.jwtSecretKey = jwtSecretKey;
+    }
+
+    protected abstract void doFilterWithJwt(JwtToken token, HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException;
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+
         final String authHeader = Optional.ofNullable(request.getHeader(HttpHeaders.AUTHORIZATION)).orElse("");
 
         // Validate authorization header
@@ -44,7 +58,6 @@ public class JwtTokenFilter extends OncePerRequestFilter {
         final String token = authHeader.substring(jwtConfig.getTokenPrefix().length()).trim();
 
         try {
-
             JwtToken jwtToken = JwtToken.from(token, jwtSecretKey);
 
             // Check token expiration
@@ -53,22 +66,12 @@ public class JwtTokenFilter extends OncePerRequestFilter {
                 return;
             }
 
-            UserAuthToken authenticationToken = new UserAuthToken(jwtToken.getUsername(), jwtToken.getAuthority());
-
-            authenticationToken.setDetails(
-                    new WebAuthenticationDetailsSource()
-                            .buildDetails(request)
-            );
-
-            val securityContext = SecurityContextHolder.getContext();
-            securityContext.setAuthentication(authenticationToken);
-
-        } catch (JwtException exception) {
-            exception.printStackTrace();
+            doFilterWithJwt(jwtToken, request, response, filterChain);
+        } catch (JwtException e) {
+            e.printStackTrace();
             throw new UnauthorizedException("The provided token cannot be trusted.");
         }
-
         filterChain.doFilter(request, response);
-
     }
+
 }
