--- src/main/java/com/aline/core/security/service/SecurityCheck.java
@@ -1,8 +1,0 @@
-package com.aline.core.security.service;
-
-import com.aline.core.dto.response.UserResponse;
-import org.springframework.security.core.Authentication;
-
-public interface SecurityCheck {
-    boolean check(UserResponse response, Authentication authentication);
-}
